Level Selector and Editor Testing
Added functionality for both Level Selection and Editor Testing.

When starting a game, the player will be able to choose which level they want to play and be loaded into the selected level.

In the level editor, players can press T to test the level. They exit the test and return to the editor when they either die or complete the level, or by pressing T again.